requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
pandas==2.1.3
tradingview-websocket==0.0.5
supabase==2.1.0
pydantic==2.5.2
postgrest-py==0.10.0
